Enhanced flag parsing 🏳️‍🌈
I set my rainbow in the cloud, and it will be a sign of a covenant between me and the earth.

```
Look at the rainbow, and praise him who made it.
It is exceedingly beautiful in its brightness.
It encircles the sky with its glorious circle.
The hands of the Most High have stretched it out.
```